+++ packages/cover-traffic-daemon/src/index.ts
@@ -136,7 +136,7 @@ export async function main(update: (State: State) => void, peerId?: PeerId) {
   log('node is running')
 
   log(node.getVersion())
-  log(await node.smartContractInfo())
+  log(node.smartContractInfo())
 
   const channels = await node.getChannelsFrom(selfAddr)
   data.setCTChannels(channels.map((c) => ({ destination: c.destination, latestQualityOf: 0, openFrom: Date.now() })))