+++ packages/utils/src/async/concurrency.spec.ts
@@ -3,38 +3,108 @@ import { oneAtATime } from './concurrency'
 
 describe('concurrency', function () {
   it('one at a time', async function () {
-    let resolveP1,
-      p1Started = false,
-      resolveP2,
-      p2Started = false,
-      out = 0
-    const p1 = () =>
-      new Promise<void>((resolve) => {
-        p1Started = true
-        out = 1
-        resolveP1 = resolve
+    const limitter = oneAtATime<void>()
+
+    const results: number[] = []
+
+    const promises = []
+
+    const CALLS = 3
+
+    for (let i = 0; i < CALLS; i++) {
+      limitter(async () => {
+        const promise = new Promise<number>((resolve) => setTimeout(resolve, 50, i))
+        promises.push(promise)
+        results.push(await promise)
       })
-    const p2 = () =>
-      new Promise<void>((resolve) => {
-        p2Started = true
-        out = 2
-        resolveP2 = resolve
+    }
+
+    await Promise.all(promises)
+
+    assert(results.length == CALLS, `must contain all results`)
+    assert(
+      results.every((value: number, index: number) => value == index),
+      `must contain the right results`
+    )
+  })
+
+  it('one at a time - restart', async function () {
+    const limitter = oneAtATime<void>()
+
+    const results: number[] = []
+
+    let promise: Promise<number>
+    limitter(async () => {
+      promise = new Promise<number>((resolve) => setTimeout(setImmediate, 20, resolve, 0))
+      results.push(await promise)
+    })
+
+    await promise
+
+    assert(results.length == 1, `must contain one result`)
+    assert(results[0] == 0)
+
+    const promises: Promise<number>[] = []
+
+    const CALLS = 3
+
+    for (let i = 0; i < CALLS; i++) {
+      limitter(async () => {
+        const promise = new Promise<number>((resolve) => setTimeout(setImmediate, 20, resolve, i + 1))
+        promises.push(promise)
+        results.push(await promise)
       })
-    const o = oneAtATime()
-    let p1prom = o(p1)
-    let p2prom = o(p2)
-    await setImmediate(() => {})
-    assert(p1Started)
-    assert(!p2Started)
-    assert(p1Started)
-    assert(!p2Started)
-    resolveP1()
-    await p1prom
-    assert(p1Started)
-    assert(p2Started)
-    assert(out == 2)
-    resolveP2()
-    await p2prom
-    assert(out == 2)
+    }
+
+    await Promise.all(promises)
+
+    // @ts-ignore
+    assert(results.length == 4, `must contain all results`)
+    assert(
+      results.every((value: number, index: number) => value == index),
+      `must contain the right results`
+    )
+  })
+
+  it('one at a time - asynchronous errors', async function () {
+    const limitter = oneAtATime<void>()
+
+    const results: number[] = []
+
+    const promises = []
+
+    const CALLS = 3
+
+    for (let i = 0; i < CALLS; i++) {
+      limitter(async () => {
+        const promise = new Promise<number>((_, reject) => setTimeout(setImmediate, 20, reject, i))
+        promises.push(promise)
+        results.push(await promise)
+      })
+    }
+
+    assert.rejects(async () => await Promise.all(promises))
+
+    assert(results.length == 0, `must contain no results`)
+  })
+
+  it('one at a time - synchronous errors', async function () {
+    const limitter = oneAtATime<void>()
+
+    const results: number[] = []
+
+    const promises = []
+
+    const CALLS = 3
+
+    for (let i = 0; i < CALLS; i++) {
+      limitter(async () => {
+        throw Error(`i`)
+      })
+    }
+
+    assert.rejects(async () => await Promise.all(promises))
+
+    assert(results.length == 0, `must contain no results`)
   })
 })